changed block::size int -> uint16_t; comments

+++ memory.h
@@ -3,6 +3,9 @@
 
 #include <vector>
 #include <string>
+#include <unordered_map>
+#include <tuple>
+#include <cstdint>
 
 struct record {
     char tconst[10];
@@ -12,37 +15,37 @@ struct record {
 
 struct block {
     std::vector<record> records;
-    int size;
+    uint16_t size;
 };
 
 class disk {
-    private:
-        std::vector<block> blocks;
-        int size;
-        int blocksize;
-        int numBlocks;
-    public:
-        disk(int blocksize);
+private:
+    std::vector<block> blocks;
+    int size;
+    int blocksize;
+    int numBlocks;
+public:
+    disk(int blocksize);
 
-        ~disk();
+    ~disk();
 
-        // prints all records in the database
-        void printitems();
+    // prints all records in the database
+    void printitems();
 
-        // create disk by reading from file
-        void readDataFromFile(std::string filePath);
+    // create disk by reading from file
+    void readDataFromFile(std::string filePath);
 
-        // returns a reference to the disk, used to build the bp tree
-        std::vector<block> *getBlock();
+    // returns a reference to the disk, used to build the bp tree
+    std::vector<block> *getBlock();
 
-        // output the number of blocks used and the size of the database
-        void reportStatistics();
+    // output the number of blocks used and the size of the database
+    void reportStatistics();
 
-        // insert new record into empty block
-        void insertRecords(std::string tconst, float averageRating, int numVotes);
+    // insert new record into empty block
+    void insertRecords(std::string tconst, float averageRating, int numVotes);
 
-        // deletes a record based on the key
-        void deleteRecord(std::string key);
+    // deletes a record based on the key
+    void deleteRecord(std::string key);
 };
 
 #endif
